requirements/runtime.txt
torch>=2.1.0
torchvision
timm
pillow
loguru
transformers>=4.45.2
huggingface-hub
sentencepiece
protobuf
accelerate>=0.26.0
zstandard
easydict
evaluate>=0.4.0
datasets>=2.16.0
jsonlines
numexpr
peft>=0.2.0
pybind11>=2.6.2
pytablewriter
rouge-score>=0.0.4
sacrebleu>=1.5.0
scikit-learn>=0.24.1
sqlitedict
tqdm-multiprocess
dill
word2number
more_itertools
qtorch
einops
qwen-vl-utils
tiktoken
librosa
human_eval
lmms-eval
imageio
bitsandbytes>=0.43.1
autoawq
